Harden Web error disclosure and upload preflight for 1.0.4
