--- .github/workflows/one-shot-104-hardening.yml
@@ -1,54 +1,0 @@
-name: Apply Ghost FTP 1.0.4 hardening
-
-on:
-  push:
-    branches:
-      - release/ghost-ftp-1.0.4-quality
-    paths:
-      - scripts/one_shot_104_hardening.py
-      - .github/workflows/one-shot-104-hardening.yml
-
-permissions:
-  contents: write
-
-jobs:
-  patch:
-    runs-on: ubuntu-24.04
-    steps:
-      - name: Checkout patch branch
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
-        with:
-          ref: release/ghost-ftp-1.0.4-quality
-          fetch-depth: 0
-
-      - name: Apply deterministic patch
-        run: python3 scripts/one_shot_104_hardening.py
-
-      - name: Validate Web hardening and release metadata
-        shell: bash
-        run: |
-          set -euo pipefail
-          php -l 'GhostFTP WEB/api.php'
-          php -l 'GhostFTP WEB/download.php'
-          php -l 'GhostFTP WEB/download-archive.php'
-          php -l 'GhostFTP WEB/preview.php'
-          php -l 'GhostFTP WEB/app/helpers.php'
-          php 'GhostFTP WEB/tests/public-error-boundary.php'
-          php 'GhostFTP WEB/tests/api-upload-preflight.php'
-          python3 scripts/audit_docs.py
-          python3 scripts/audit_web.py
-          python3 scripts/audit_version.py
-          python3 scripts/audit_release.py
-
-      - name: Commit clean validated tree
-        shell: bash
-        run: |
-          set -euo pipefail
-          rm -f scripts/one_shot_104_hardening.py .github/workflows/one-shot-104-hardening.yml
-          git config user.name 'Brendigo'
-          git config user.email 'info@brendigo.com'
-          git add -A
-          git diff --cached --check
-          git status --short
-          git commit -m 'Harden Web error disclosure and upload preflight for 1.0.4'
-          git push origin HEAD:release/ghost-ftp-1.0.4-quality
