+++ GhostFTP WEB/composer.json
@@ -1,6 +1,6 @@
 {
   "name": "brendigo/ghost-ftp-web",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "description": "Ghost FTP hardened multi-user shared-hosting FTP/FTPS/SFTP PWA client",
   "type": "project",
   "require": {
