+++ README.md
@@ -2,7 +2,7 @@
 
 **Ghost FTP** is a privacy-focused FTP, FTPS and SFTP client for Windows, Linux, macOS, Android, iOS and the web. The project prioritizes predictable file-transfer behavior, explicit trust boundaries, reproducible releases and zero application telemetry.
 
-Current Ghost FTP version: **1.0.3**
+Current Ghost FTP version: **1.0.4**
 
 ## Identity
 
@@ -28,7 +28,7 @@ Canonical locations:
 
 Remote paths, local paths, temporary files, credentials and release provenance are treated as security boundaries. Regression coverage includes traversal rejection, host-key policy, transfer cleanup, configuration durability, process lifecycle, runtime-secret handling and fail-closed recovery behavior.
 
-Ghost FTP 1.0.3 tightens Web/PWA fail-closed behavior around SFTP trust, inline editor staging, SFTP key material and batch mutations. SFTP profiles require a canonical pinned SHA-256 host fingerprint before persistence, editor/new-file writes share one 4 MiB boundary with complete local staging checks, SFTP uploads verify remote size when available, and malformed/duplicate batch rename input is rejected before remote mutation.
+Ghost FTP 1.0.4 extends Web/PWA fail-closed behavior to public error handling and multi-file uploads. Unexpected PHP/extension failures no longer expose raw internal messages to clients, and complete upload request metadata/path validation now finishes before the first remote mutation. The 1.0.3 SFTP trust, staging, key-permission and batch-mutation protections remain enforced by regression audits.
 
 The web application uses CSRF protection, strict sessions, secure cookies, security headers, rate limiting and `noindex` directives. Production CI runs Go formatting, unit tests, the race detector, `go vet`, repository/version/documentation/release/security/privacy/web audits, PHP/JavaScript validation and native platform builds.
 
@@ -42,7 +42,7 @@ Usernames such as `account@domain` are supported. FTP directory listings retain 
 
 ## Releases
 
-Ghost FTP releases use `ghostftp-vX.Y.Z`. The current product line starts at `ghostftp-v1.0.0` and advances sequentially through patch releases such as `ghostftp-v1.0.1`, `ghostftp-v1.0.2` and `ghostftp-v1.0.3`.
+Ghost FTP releases use `ghostftp-vX.Y.Z`. The current product line starts at `ghostftp-v1.0.0` and advances sequentially through patch releases such as `ghostftp-v1.0.1`, `ghostftp-v1.0.2`, `ghostftp-v1.0.3` and `ghostftp-v1.0.4`.
 
 | Platform | Public package |
 | --- | --- |
@@ -63,7 +63,7 @@ Windows portable binaries are additionally published as the GitHub Packages NuGe
 
 ## Versioning policy
 
-Ghost FTP starts at **1.0.0** and follows Semantic Versioning. Patch releases advance sequentially (`1.0.0` → `1.0.1` → `1.0.2` → `1.0.3`), backward-compatible feature releases advance the minor version, and breaking compatibility requires a major-version change.
+Ghost FTP starts at **1.0.0** and follows Semantic Versioning. Patch releases advance sequentially (`1.0.0` → `1.0.1` → `1.0.2` → `1.0.3` → `1.0.4`), backward-compatible feature releases advance the minor version, and breaking compatibility requires a major-version change.
 
 Published historical tags remain immutable for provenance. Current Ghost FTP releases use the dedicated `ghostftp-vX.Y.Z` namespace so they do not collide with historical generic `vX.Y.Z` tags.
